+++ components/marketing/desk/index.tsx
@@ -1,10 +1,10 @@
-export { DeskKicker, DeskVerified, SourceLinks, DeskCtas } from "./DeskChrome";
+export { DeskKicker, DeskVerified, SourceLinks, DeskCtas, LeadTakeaway, PrimarySource } from "./DeskChrome";
 export { StatusPill, type DeskChromeStatus } from "./StatusChip";
 export { ContestedRateStrip } from "./ContestedRateStrip";
 export { TrackerBoard } from "./TrackerBoard";
+export { Receipts } from "./Receipts";
 export {
   CONTESTED_RATE_SLOTS,
-  findContestedMarker,
   resolveContestedSlots,
   type ContestedSlot,
   type ContestedSlotId,